install window size hook, create two layouts
