+++ squat/package.json
@@ -11,6 +11,7 @@
     "@capacitor/status-bar": "4.0.1",
     "@ionic/core": "^6.3.0",
     "@ionic/react": "^6.3.0",
+    "@react-hook/window-size": "^3.1.1",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^14.4.3",
